+++ src/hotelidealuno/HuespedData.java
@@ -38,7 +38,7 @@ public class HuespedData {
             statement.setInt(2, huesped.getDni());
             statement.setString(3, huesped.getDomicilio());
             statement.setString(4,huesped.getCorreo());
-            statement.setInt(5, huesped.getCelular());
+            statement.setLong(5, huesped.getCelular());
             
             statement.executeUpdate();
             
@@ -73,7 +73,7 @@ public class HuespedData {
                 huesped.setDni(resultSet.getInt("dni"));
                 huesped.setDomicilio(resultSet.getString("domicilio"));
                 huesped.setCorreo(resultSet.getString("correo"));
-                huesped.setCelular(resultSet.getInt("celular"));
+                huesped.setCelular(resultSet.getLong("celular"));
     
                 huespedes.add(huesped);
             }      
@@ -119,7 +119,7 @@ public class HuespedData {
             //statement.setDate(2, Date.valueOf(huesped.getFecNac()));
             statement.setString(3, huesped.getDomicilio());
             statement.setString(4, huesped.getCorreo());
-            statement.setInt(5, huesped.getCelular());
+            statement.setLong(5, huesped.getCelular());
             statement.setInt(6, huesped.getId());
             statement.executeUpdate();
     
@@ -152,7 +152,7 @@ public class HuespedData {
                 huesped.setDni(resultSet.getInt("dni"));
                 huesped.setDomicilio(resultSet.getString("domicilio"));
                 huesped.setCorreo(resultSet.getString("correo"));
-                huesped.setCelular(resultSet.getInt("celular"));
+                huesped.setCelular(resultSet.getLong("celular"));
                 
             }      
             statement.close();
